Website/backend/requirements.txt
gradio==4.12.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.35.0
Pillow>=10.0.0
